% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costPer100.r
\name{costPer100}
\alias{costPer100}
\title{calculate cost per 100 miles for each car}
\usage{
costPer100(df)
}
\arguments{
\item{df}{mtcars dataset by default}
}
\value{
dataframe with two columns - one with car names and the other with cost to drive 100 miles
}
\description{
Based on the miles per gallon and current gallon cost - 3$ - the function returns sorted cost list for each car
}
