% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showBrands.r
\name{showBrands}
\alias{showBrands}
\title{Show all car brands inside mtcars}
\usage{
showBrands(df)
}
\arguments{
\item{df}{mtcars dataset by default}
}
\value{
dataframe with list of uniqe brands based on car name
}
\description{
Based on the car names the function returns all unique brand names for the cars in the dataset
}
