% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightToKg.r
\name{weightToKg}
\alias{weightToKg}
\title{Change car weight from lbs to kg}
\usage{
weightToKg(df)
}
\arguments{
\item{df}{mtcars dataset by default}
}
\value{
dataframes with two columns - one with carnames and the other with weight in kgs
}
\description{
This function allows you to change the unit to kg for each car weight
}
